finalizando projeto

+++ src/routes.js
@@ -29,12 +29,29 @@ export const routes = [
     handler: (req, res) => {
       const { name, description } = req.body;
 
+      if (!name) {
+        return res.writeHead(400).end(
+          JSON.stringify({
+            error: "Name is required field.",
+          })
+        );
+      }
+
+      if (!description) {
+        return res.writeHead(400).end(
+          JSON.stringify({
+            error: "Description is required field.",
+          })
+        );
+      }
+
       const task = {
         id: randomUUID(),
         name,
         description,
         completed_at: null,
         created_at: new Date(),
+        updated_at: new Date(),
       };
 
       database.insert("task", task);
@@ -48,6 +65,30 @@ export const routes = [
     handler: (req, res) => {
       const { id } = req.params;
       const { name, description } = req.body;
+
+      if (!id) {
+        return res.writeHead(404).end(
+          JSON.stringify({
+            error: "Task not found.",
+          })
+        );
+      }
+
+      if (!name) {
+        return res.writeHead(400).end(
+          JSON.stringify({
+            error: "Name is required field.",
+          })
+        );
+      }
+
+      if (!description) {
+        return res.writeHead(400).end(
+          JSON.stringify({
+            error: "Description is required field.",
+          })
+        );
+      }
 
       database.update("task", id, {
         name,
@@ -65,6 +106,14 @@ export const routes = [
       const { id } = req.params;
       const { completed_at } = req.body;
 
+      if (!id) {
+        return res.writeHead(404).end(
+          JSON.stringify({
+            error: "Task not found.",
+          })
+        );
+      }
+
       database.update("task", id, {
         completed_at,
       });
@@ -77,6 +126,14 @@ export const routes = [
     path: buildRoutePath("/task/:id"),
     handler: (req, res) => {
       const { id } = req.params;
+
+      if (!id) {
+        return res.writeHead(404).end(
+          JSON.stringify({
+            error: "Task not found.",
+          })
+        );
+      }
 
       database.delete("task", id);
